Use 'requests'; Multi-stage build; Pagination
Replace the use of 'httplib' with 'requests' to mitigate a vulnerability
with httplib on older versions of Python. requests is ultimately a bit
easier to work with, anyway.

This updates the Docker build for multi-stage images with a virtual
environment to run in.

Additionally, the Docker Hub results with pagination are now parsed
correctly.